refactor: apply UNIX philosophy to task management components
- Extract duplicate STATUSES/TYPES/STATUS_CHIP/TYPE_CHIP constants from
  TaskRow and NewTaskRow into shared task-options.ts
- Move pure date helpers (getDateRange, formatDate, isWeekend) out of
  TaskTimelineView into task-utils.ts
- Extract duplicated color swatch JSX into a local ColorPicker component
  in TeamManager, replacing two identical 9-line blocks
